* Use const mask in LASClassification. * Added a couple of new test cases for LASClassification. * Updated build/msvc90/las2ogr project.
--HG--
extra : convert_revision : svn%3A1766ff46-f334-0410-ab20-d63176f87757/trunk%401222

+++ include/liblas/lasclassification.hpp
@@ -130,7 +130,7 @@ public:
     uint8_t GetClass() const
     {
         bitset_type bits(m_flags);
-        bitset_type const mask(class_table_size - 1);
+        bitset_type const mask(static_cast<unsigned long>(class_table_size) - 1);
         bits &= mask;
 
         uint8_t const index = static_cast<uint8_t>(bits.to_ulong());
